feat: add role permission on bill and payment and academic master data

+++ app/Http/Controllers/Admin/AcademicYearController.php
@@ -6,9 +6,10 @@ use App\Models\AcademicYear;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use Illuminate\Support\Facades\DB;
-use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\AcademicYearRequest;
 use Illuminate\Support\Facades\Log;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use App\Http\Requests\Admin\AcademicYearRequest;
 
 class AcademicYearController extends Controller
 {
@@ -17,8 +18,12 @@ class AcademicYearController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Tahun Ajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
+
         if (request()->ajax()) {
-            $data = AcademicYear::latest()->get();
+            $data = AcademicYear::latest();
             return DataTables::of($data)
                 ->addColumn('status', function ($data) {
                     return $data->is_active ? '<span class="badge badge-success">Aktif</span>'
@@ -29,9 +34,9 @@ class AcademicYearController extends Controller
                     $actionEdit = route('academic-year.edit', $data->id);
                     $actionDelete = route('academic-year.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.status', ['action' => $actionStatus, 'status' => $data->is_active, 'id' => $data->id]) .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.status', ['action' => $actionStatus, 'status' => $data->is_active, 'id' => $data->id, 'name' => 'Tahun Ajaran']) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Tahun Ajaran']) .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Tahun Ajaran']) .
                         "</div>";
                 })
                 ->rawColumns(['action', 'status'])
@@ -45,6 +50,9 @@ class AcademicYearController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Tahun Ajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.academic-year.create-edit');
     }
 
@@ -53,6 +61,9 @@ class AcademicYearController extends Controller
      */
     public function store(AcademicYearRequest $request)
     {
+        if (!Auth::user()->can('Create Tahun Ajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         AcademicYear::create($request->validated());
         return redirect()->route('academic-year.index')->with('success', 'Tahun Akademik berhasil ditambahkan');
     }
@@ -70,6 +81,9 @@ class AcademicYearController extends Controller
      */
     public function edit(AcademicYear $academicYear)
     {
+        if (!Auth::user()->can('Edit Tahun Ajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.academic-year.create-edit', compact('academicYear'));
     }
 
@@ -78,6 +92,9 @@ class AcademicYearController extends Controller
      */
     public function update(AcademicYearRequest $request, AcademicYear $academicYear)
     {
+        if (!Auth::user()->can('Edit Tahun Ajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $academicYear->update($request->validated());
         return redirect()->route('academic-year.index')->with('success', 'Tahun Akademik berhasil diperbarui');
     }
@@ -87,12 +104,18 @@ class AcademicYearController extends Controller
      */
     public function destroy(AcademicYear $academicYear)
     {
+        if (!Auth::user()->can('Delete Tahun Ajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $academicYear->delete();
         return redirect()->route('academic-year.index')->with('success', 'Tahun Akademik berhasil dihapus');
     }
 
     public function status(string $id)
     {
+        if (!Auth::user()->can('Edit Tahun Ajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         try {
             DB::beginTransaction();
