+++ app/Http/Controllers/Admin/StudyController.php
@@ -6,6 +6,7 @@ use App\Models\Study;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Admin\StudyRequest;
 
 class StudyController extends Controller
@@ -15,15 +16,18 @@ class StudyController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Mata Pelajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         if (request()->ajax()) {
-            $data = Study::latest()->get();
+            $data = Study::latest();
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
                     $actionEdit = route('study.edit', $data->id);
                     $actionDelete = route('study.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Mata Pelajaran']) .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Mata Pelajaran']) .
                         "</div>";
                 })
                 ->rawColumns(['action'])
@@ -37,6 +41,9 @@ class StudyController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Mata Pelajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.study.create-edit');
     }
 
@@ -45,6 +52,9 @@ class StudyController extends Controller
      */
     public function store(StudyRequest $request)
     {
+        if (!Auth::user()->can('Create Mata Pelajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $data = $request->validated();
         Study::create($data);
         return redirect()->route('study.index')->with('success', 'Mata Pelajaran berhasil ditambahkan');
@@ -63,6 +73,9 @@ class StudyController extends Controller
      */
     public function edit(Study $study)
     {
+        if (!Auth::user()->can('Edit Mata Pelajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.study.create-edit', compact('study'));
     }
 
@@ -71,6 +84,9 @@ class StudyController extends Controller
      */
     public function update(StudyRequest $request, Study $study)
     {
+        if (!Auth::user()->can('Edit Mata Pelajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $data = $request->validated();
         $study->update($data);
         return redirect()->route('study.index')->with('success', 'Mata Pelajaran berhasil diubah');
@@ -81,6 +97,9 @@ class StudyController extends Controller
      */
     public function destroy(Study $study)
     {
+        if (!Auth::user()->can('Delete Mata Pelajaran')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $study->delete();
         return redirect()->route('study.index')->with('success', 'Mata Pelajaran berhasil dihapus');
     }
